fix BalanceAmount

+++ src/pages/address.tsx
@@ -58,7 +58,7 @@ export default function Address() {
                   <Heading size="sm" mt={index === 0 ? 0 : 4}>
                     {key}
                   </Heading>
-                  <BalanceAmount token={key} value={value as bigint} />
+                  <BalanceAmount token={key} value={value} />
                 </Fragment>
               ))}
             </Box>
